third_party/blink/web_tests/external/wpt/css/css-conditional/container-queries/query-container-name-expected.txt

This is a testharness.js-based test.
[FAIL] match closest named container
  assert_equals: expected "yes" but got "no"
[FAIL] match ancestor named container
  assert_equals: expected "yes" but got "no"
Harness: the test ran to completion.
